[tools] Allow operator to specify whether or not to drop DB on creation

+++ ndk/relay/event_repo/postgres_event_repo.py
@@ -89,15 +89,18 @@ class PostgresEventRepo(event_repo.EventRepo):
         return engine
 
     @classmethod
-    async def create(cls, host, port, user, password, database):
+    async def create(cls, host, port, user, password, database, drop_db=False):
         engine = await cls.create_engine(host, port, user, password, database)
 
-        async with engine.begin() as conn:  # transaction
-            await conn.run_sync(
-                METADATA.drop_all,
-                tables=[EVENT_TAGS_TABLE, TAGS_TABLE, EVENTS_TABLE],
-                checkfirst=True,
-            )
+        async with engine.begin() as conn:
+            if drop_db:
+                logger.info("Dropping DB")
+                await conn.run_sync(
+                    METADATA.drop_all,
+                    tables=[EVENT_TAGS_TABLE, TAGS_TABLE, EVENTS_TABLE],
+                    checkfirst=True,
+                )
+
             await conn.run_sync(METADATA.create_all)
 
         logger.info("Database initialized")
